% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearReg.R
\name{linearPred}
\alias{linearPred}
\title{Predict outputs of the given x values using the given model}
\usage{
linearPred(model, x)
}
\arguments{
\item{model}{input model}

\item{x}{input values}
}
\description{
Predict outputs of the given x values using the given model
}
