% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearreg.R
\name{linearreg}
\alias{linearreg}
\title{Linearreg}
\usage{
linearreg(y, X, prednames)
}
\arguments{
\item{y}{vector of outcome values}

\item{X}{matrix whose columns are predictor values, ordered like y. If missing fits intercept-only}

\item{prednames}{optional names for predictors, in order of columns of X}
}
\description{
Fits a linear regression model and does inference
}
\examples{
y <- c(1,2,3,4,5,6,7,8,9,10)
x1 <- c(10,8,9,6,5,9,3,2,1,3)
x2 <- c(6,4,8,6,2,9,0,1,4,6)
prednames <- c("Speed", "Weight")
linearreg(y, cbind(x1, x2))

}
