% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{package}
\name{linreg}
\alias{linreg}
\title{Linreg}
\arguments{
\item{formula}{Formula for the linear regression.}

\item{data}{A dataframe with observations.}
}
\value{
This package returns many statistics.
\describe{
 \item{Regression.Coefficient}{Regressions coefficients}
 \item{Fitted.Values}{The fitted values}
 \item{Residuals}{The residuals}
 \item{df}{The degrees of freedom}
 \item{Residual.Variance}{The residual variance}
 \item{Variance.B}{The variance of the regression coefficients}
 \item{T.Values}{The t-values for each coefficient}
 \item{P.Values}{The p-values for each coefficient}
 \item{call}{The arguments used to call the function}
}
}
\section{Fields}{

\describe{
\item{\code{Several}}{statistics.}
}}

