% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearReg.R
\name{summary_linearReg}
\alias{summary_linearReg}
\title{Print the summary of the linear model}
\usage{
summary_linearReg(lm)
}
\arguments{
\item{lm}{The object returned by the function linearReg.}
}
\description{
This function print the attributes of the linear regression model, including coefficents, and their
stand error, t statistics and p values. As for the model, it will print out the residual standard error
and R square (both raw and adjusted). At last, it will print the F statistics and its degree of freedom.
}
