% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearReg.R
\name{linearReg}
\alias{linearReg}
\title{Applies linear regression and finds the coefficient values}
\usage{
linearReg(x, y)
}
\arguments{
\item{x}{input values}

\item{y}{output values}
}
\description{
Applies linear regression and finds the coefficient values
}
