% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearReg.R
\name{linearReg}
\alias{linearReg}
\title{Basic linear regression}
\usage{
linearReg(Y, X, has_intercept = TRUE)
}
\arguments{
\item{Y}{A vector or matrix (one column only) of outcome values.}

\item{X}{A vector or matrix (one or multiple columns) of predictors.}

\item{has_intercept}{A boolean variable determines whether the model includes a intercept.}
}
\value{
A list of attributes of the linear model.
}
\description{
This function is used to build linear models. It can be used to carry out simple linear regression (SLR),
multiple linear regression (MLR).
}
