% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg}
\alias{linreg}
\title{Linreg}
\arguments{
\item{formula}{Formula for the linear regression.}

\item{data}{A dataframe with observations.}
}
\description{
Linear regression package in R.
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{regression.coefficients}}{matrix.}

\item{\code{fitted.values}}{matrix.}

\item{\code{residuals}}{matrix.}

\item{\code{df}}{numeric.}

\item{\code{residual.variance}}{numeric.}

\item{\code{variance.regression.coefficients}}{matrix.}

\item{\code{t.vals}}{matrix.}

\item{\code{p.vals}}{matrix.}

\item{\code{call}}{character.}

\item{\code{sq.coeff}}{numeric.}
}}

